DE_answers/requirements.txt
boto3==1.28.59
botocore==1.31.59
confluent-kafka==2.2.0
iniconfig==2.0.0
jmespath==1.0.1
numpy==1.26.0
packaging==23.2
pandas==2.1.1
pluggy==1.3.0
psycopg2-binary==2.9.8
pytest==7.4.2
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
s3transfer==0.7.0
six==1.16.0
tzdata==2023.3
urllib3==1.26.17
